refactor: saga test for send swap
